+++ gulp/patternlab.js
@@ -1,0 +1,134 @@
+/******************************************************
+ * EXPORTS PATTERN LAB TASKS
+ *
+ * Note that substantial chunks of this file were
+ * originally copied from Pattern Lab's Node Gulp
+ * Edition gulpfile:
+ * https://github.com/pattern-lab/edition-node-gulp/blob/master/gulpfile.js
+ *
+******************************************************/
+
+const gulp = require('gulp');
+const path = require('path');
+const argv = require('minimist')(process.argv.slice(2));
+
+const browserSync = require('./browsersync.js');
+
+const taskNamePrefix = 'patternlab:';
+
+//read all paths from our namespaced config file
+const config = require('../patternlab-config.json');
+const patternlab = require('@pattern-lab/patternlab-node')(config);
+
+
+/******************************************************
+ * CLI TASKS
+******************************************************/
+
+function plVersionTask (done) {
+  patternlab.version();
+  done();
+};
+plVersionTask.displayName = taskNamePrefix + 'version';
+plVersionTask.description = 'eturn the version of patternlab-node you have installed';
+
+
+function plHelpTask (done) {
+  patternlab.help();
+  done();
+};
+plHelpTask.displayName = taskNamePrefix + 'help';
+plHelpTask.description = 'Get more information about patternlab-node, pattern lab in general, and where to report issues.';
+
+
+function plPatternsOnlyTask (done) {
+  patternlab.patternsonly(done, getConfiguredCleanOption());
+};
+plPatternsOnlyTask.displayName = taskNamePrefix + 'patternsonly';
+plPatternsOnlyTask.description = 'Compiles the patterns only, outputting to config.paths.public';
+
+
+function plListStarterKitsTask (done) {
+  patternlab.liststarterkits();
+  done();
+};
+plListStarterKitsTask.displayName = taskNamePrefix + 'liststarterkits';
+plListStarterKitsTask.description = 'Returns a url with the list of available starterkits hosted on the Pattern Lab organization Github account';
+
+
+function plLoadStarterKitsTask (done) {
+  patternlab.loadstarterkit(argv.kit, argv.clean);
+  done();
+};
+plLoadStarterKitsTask.displayName = taskNamePrefix + 'loadstarterkit';
+plLoadStarterKitsTask.description = 'Load a starterkit into config.paths.source/*';
+
+
+function plInstallPluginTask (done) {
+  patternlab.installplugin(argv.plugin);
+  done();
+};
+plInstallPluginTask.displayName = taskNamePrefix + 'installplugin';
+
+
+/******************************************************
+ * BUILD TASK
+******************************************************/
+
+function plBuildTask(watch) {
+  return patternlab.build({
+    watch: watch || argv.watch,
+    cleanPublic: config.cleanPublic
+  }).then(() => {
+    // do something else when this promise resolves
+  });
+}
+plBuildTask.displayName = taskNamePrefix + 'build';
+plBuildTask.description = 'Compiles the patterns and frontend, outputting to config.paths.public';
+
+
+/******************************************************
+ * WATCH TASK
+******************************************************/
+
+function plWatchTask() {
+  return plBuildTask(true);
+}
+plWatchTask.displayName = taskNamePrefix + 'watch';
+plWatchTask.description = 'Builds the styleguide and starts watching styleguide source files.';
+
+
+/******************************************************
+ * SERVE TASK
+******************************************************/
+
+function plServeTask() {
+  return patternlab.serve({
+    cleanPublic: config.cleanPublic
+  }).then(() => {
+    // do something else when this promise resolves
+  });
+}
+plServeTask.displayName = taskNamePrefix + 'serve';
+plServeTask.description = 'Builds styleguide and launches Pattern Lab\'s built-in server.';
+
+
+
+module.exports = {
+  // CLI
+  plVersionTask,
+  plHelpTask,
+  plPatternsOnlyTask,
+  plListStarterKitsTask,
+  plLoadStarterKitsTask,
+  plInstallPluginTask,
+
+  // Build
+  plBuildTask,
+
+  // Watch
+  plWatchTask,
+
+  // Serve
+  plServeTask,
+}