+++ CHANGELOG.md
@@ -1,14 +1,31 @@
 # Changelog
-All notable changes to this project will be documented in this file.
+All notable changes to the [`gravity-ui-sass` project](./README.md) will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
+
+
+## [0.2.0] - 2018-01-19
 ### Added
-- CHANGELOG
+- NVM config
+- Basic font styles
+- Normalize and modularscale SASS libs
+- `gravity-ui-sass` is now an [SASS Eyeglass](https://github.com/sass-eyeglass/eyeglass) module
+- Package now exposes paths to CSS and SASS to cosnumers via its main entry point (`index.js`)
 
 ### Changed
-- fixed initial semver
-- updated stylelint rules and dependencies
-- updated README.md with information about versioning and release process
+- Improved Stylelint config
+- Split gulpfile.js into modules
+
+
+## [0.1.0] - 2018-01-16
+### Added
+- Initial SASS scaffolding
+- Initial README, CHANGELOG, etc. docs
+- [Stylelint](https://stylelint.io/) rules
+- Gulp setup for running SASS compilation and [Pattern Lab](http://patternlab.io/)
+- Initial set of UI patterns for Pattern Lab
+- Initial [Gravy](https://github.com/buildit/gravy) integration